# Improve cuBLAS performance by dequantizing on the GPU