For me this makes cuBLAS about twice as fast with quantized models.

Perplexity seconds per pass
| Model | PR | Master |
| --- | --- | --- |
| q4_0 | 5.05 | 8.62|
| q4_1 | 5.37 | 8.59 |
| q4_2 | 4.99 | 10.76 |


Prompt eval time with 7B q4_0 (bs=512)
```
cuBLAS (PR):     prompt eval time =  7840.48 ms /   631 tokens (   12.43 ms per token)
cuBLAS (Master): prompt eval time = 15457.33 ms /   631 tokens (   24.50 ms per token)
OpenBLAS:        prompt eval time = 34856.06 ms /   631 tokens (   55.24 ms per token)
No BLAS:         prompt eval time = 43549.67 ms /   631 tokens (   69.02 ms per token)
```

13B q4_0
```
cuBLAS (PR):     prompt eval time = 13826.48 ms /   631 tokens (   21.91 ms per token)
cuBLAS (Master): prompt eval time = 27987.82 ms /   631 tokens (   44.35 ms per token)
OpenBLAS:        prompt eval time = 61476.58 ms /   631 tokens (   97.43 ms per token)
No BLAS:         prompt eval time = 81645.43 ms /   631 tokens (  129.39 ms per token)
```