@slaren 2023-04-19
I just added some prompt eval times for 7B q4_0.