@glinscott 2023-04-19
Wow, this is a game changer!  Interestingly, 16 threads and 8 threads seems to be same speed now.  Only uses ~600MB of GPU RAM (RTX 3080), and GPU utilization 65% or so.  Amazing work!

All tests run with: `$ ./perplexity -m models/7B/ggml-model-q4_0.bin -f wiki.test.raw -c 512 -t <N>`

w/ cuBLAS (this PR):
```
1 thread  - 5.12 seconds per pass - ETA 0.93 hours
2 threads - 3.82 seconds per pass - ETA 0.70 hours
4 threads - 3.22 seconds per pass - ETA 0.59 hours
8 threads - 2.93 seconds per pass - ETA 0.53 hours
16 threads - 2.82 seconds per pass - ETA 0.51 hours
```

Without CUDA (8 threads):
```
8 threads  - 16.57 seconds per pass - ETA 3.01 hours
16 threads - 11.46 seconds per pass - ETA 2.08 hours
```