@glinscott 2023-04-19
Even more incredible, this allows me to run the full 65B model on a machine with 32GB of RAM quite quickly!
```
7B - 2.93 seconds per pass - ETA 0.53 hours
13B - 4.86 seconds per pass - ETA 0.88 hours
30B - 10.99 seconds per pass - ETA 2.00 hours
65B - 37.98 seconds per pass - ETA 6.91 hours
```

Btw, for comparison, from last month 7B was at `24.58 seconds per pass - ETA 4.47 hours`!