@slaren 2023-04-19
Nice! Don't use this just yet to run perplexity computations though, I found a synchronization issue that may cause  inaccurate results. Should be fixed in the last commit though, I am running a full perplexity test and if it looks good it will be ready to merge.