@glinscott 2023-04-19
> > Even more incredible, this allows me to run the full 65B model on a machine with 32GB of RAM quite quickly!
> 
> wait... how does that work. are you not supposed to need ~60gigs of ram for 65B ?

I'm using mmap mode, so it has to go to disk to read parts of the model in as it's going.  That was brutally slow previously, but the overlap with running things on GPU seems to make it feasible now.
```
cuBLAS - 37.98 seconds per pass - ETA 6.91 hours
CPU - 109.73 seconds per pass - ETA 19.96 hours
```

Actually, even on CPU it's much better than it used to be.  Everyone doing amazing work here :).