@slaren 2023-04-19
On a side node, should we increase the default batch size when ggml is built with BLAS support? Would it make it easier to use.