@SlyEcho 2023-04-20
I ported this to HIP and hipBLAS.

The dequant on the device is nothing earth-shattering. Still dominated by the memcpy. But I have an older card and only PCIe 3.0.

![image](https://user-images.githubusercontent.com/795193/233228126-91fac2b2-1ac4-46a6-9486-ddd81b649acd.png)