@Dampfinchen 2023-04-20
Hello there. I'm trying to build it using the make LLAMA_CUBLAS=1 command with Windows and WIN64DevKit. However, even though I have CUDA Toolkit installed and changed the paths for -L and -I in the makefile accordingly, it still misses the following libaries:
-lcublas_static
 -lculibos
 -lcublasLt_static
and -ldl:.

Where can I get them? I would appreciate some help getting this to work. Thank you!